Continue troubleshooting...

+++ os-setup/ci-postboot-setup.sh
@@ -1,7 +1,13 @@
 #!/bin/bash -eux
 # This script must be run as root.
 
+if [ ! -f /var/lib/os-setup/success ]; then
+  echo "Error: the OS setup scripts did not complete successfully during the previous boot!"
+  return 1
+fi
+
 rm -rf /usr/lib/os-setup
+rm -rf /var/lib/os-setup
 
 systemctl disable ci-boot-setup.service
 rm /usr/lib/systemd/system/ci-boot-setup.service
